Remove repeating logic from parse_operator_key() and fix bug in apply_operator() in Circuit
The bug in apply_operator() was that the variable "key" was being passed to construct_operator() instead of the variable "preprocessed_key".

This caused the tests for partial superposition measurement, single qubit gate at the end, and single qubit gate at the start, to fail when the repeating logic was removed, but all other tests passed.